% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{APA}
\alias{APA}
\title{Creating an APA citation}
\usage{
APA(name, key)
}
\arguments{
\item{name}{The name of the article}

\item{key}{The API key of the user}
}
\value{
A character string of APA citation
}
\description{
Given the name of the article and the user's API key, this
function returns a character string of APA citation for that article.
}
