% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff2bibentry.R
\name{cff2bibentry}
\alias{cff2bibentry}
\title{cff2bibentry}
\usage{
cff2bibentry(x)
}
\arguments{
\item{x}{path to citation file or read in citation in citation file format}
}
\description{
Converter citation file format to bibentry
}
\author{
Jan Philipp Dietrich
}
