% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff2md.R
\name{cff2md}
\alias{cff2md}
\title{cff2md}
\usage{
cff2md(x)
}
\arguments{
\item{x}{path to citation file or read in citation in citation file format}
}
\description{
Creates a summary of the CFF content in markdown format. This
can be useful for instance to create software reports.
}
\author{
Jan Philipp Dietrich
}
