% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff2zenodo.R
\name{cff2zenodo}
\alias{cff2zenodo}
\title{cff2zenodo}
\usage{
cff2zenodo(x)
}
\arguments{
\item{x}{path to citation file or read in citation in citation file format. If a path is provided the .zenodo.json information will be written
directly to a .zenodo.json file in the corresponding folder. Otherwise the metadata will be just returned by the function.}
}
\value{
The metadata information that should be provided to Zenodo in JSON format.
}
\description{
Converter citation file format to .zenodo.json metadata file for proper
metadata input to Zenodo.
}
\author{
Jan Philipp Dietrich
}
