% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffReport.R
\name{cffReport}
\alias{cffReport}
\title{cffReport}
\usage{
cffReport(repos, file = NULL, sortBy = "title")
}
\arguments{
\item{repos}{character vector of repository urls or links to CITATION.cff
files to be scanned}

\item{file}{character string naming a file the report should be written to. If
NULL the report will be directly returned by the function.}

\item{sortBy}{category by which the report should sort its entries}
}
\value{
the report in markdown format
}
\description{
Creates a report in markdown format for a provided list of repositories
and or CITATION.cff files.
}
\author{
Jan Philipp Dietrich
}
