% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation-package.R
\docType{package}
\name{citation-package}
\alias{citation}
\alias{citation-package}
\title{citation: Software Citation Tools}
\description{
A collection of functions to extract citation information from 'R' packages and to deal with files in 'citation file format' (\url{https://citation-file-format.github.io/}), extending the functionality already provided by the citation() function in the 'utils' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/citation}
  \item \doi{10.5281/zenodo.3813429}
  \item Report bugs at \url{https://github.com/pik-piam/citation/issues}
}

}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}

Authors:
\itemize{
  \item Waldir Leoncio \email{w.l.netto@medisin.uio.no}
}

}
