% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation2zenodo.R
\name{citation2zenodo}
\alias{citation2zenodo}
\title{citation2zenodo}
\usage{
citation2zenodo(cit, folder = NULL)
}
\arguments{
\item{cit}{an object of class citation as returned for instance by the function \code{\link{citation}},
or the name of a R package to be used as source.}

\item{folder}{a folder the file should be written to. If NULL the content is instead printed.}
}
\description{
Converts an object of class citation into a .zenodo.json file. This file is an
(officially unsupported) method to provide ZENODO with metadata information
about the package and will overwrite the information ZENODO would otherwise 
auto-generate from the repository.
}
\examples{

cit <- citation_doi("citation")
citation2zenodo(cit)

}
\author{
Jan Philipp Dietrich
}
