% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citation_data}
\alias{citation_data}
\title{This is a help file for the 1 exported data set, Metadata for APA citation}
\format{
A data frame with 1 rows and 4 columns
\describe{
\item{authors}{author(s) name}
\item{year}{Year article was published}
\item{title}{Title of the article}
\item{publication}{Additional publication material including Journal title, Volume of journal, Issue number of journal and Page range of article.
}
}
\source{
\url{https://link.springer.com/article/10.1007/s11136-004-0612-6}
}
\usage{
citation_data
}
\description{
This is a help file for the 1 exported data set, Metadata for APA citation
}
\keyword{datasets}
