% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_doi.R
\name{citation_doi}
\alias{citation_doi}
\title{citation_doi}
\usage{
citation_doi(package)
}
\arguments{
\item{package}{a character string with the name of a single package.}
}
\description{
Extension of the standard citation function which also extracts
DOI information from the URL field in the DESCRIPTION file,
if available.
}
\examples{

citation_doi("citation")

}
\author{
Jan Philipp Dietrich
}
