% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package2zenodo.R
\name{package2zenodo}
\alias{package2zenodo}
\title{package2zenodo}
\usage{
package2zenodo(package = ".")
}
\arguments{
\item{package}{either the path to the main folder of a package (containing a DESCRIPTION file)
or the name of the package. If a path is provided the .zenodo.json file will be directly added
to the folder. Otherwise the content of such a file is just returned.}
}
\value{
The metadata information that should be provided to Zenodo in JSON format.
}
\description{
Creates a .zenodo.json for a R package. This file is an
(officially unsupported) method to provide ZENODO with metadata information
about the package and will overwrite the information ZENODO would otherwise
auto-generate from the repository.
}
\examples{

package2zenodo("citation")
}
\author{
Jan Philipp Dietrich
}
