% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cff.R
\name{read_cff}
\alias{read_cff}
\title{read_cff}
\usage{
read_cff(file)
}
\arguments{
\item{file}{path to the citation file}
}
\description{
Read a citation file format  file (https://citation-file-format.github.io/)
}
\author{
Jan Philipp Dietrich
}
