% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanRepos.R
\name{scanRepos}
\alias{scanRepos}
\title{scan repositories}
\usage{
scanRepos(repos)
}
\arguments{
\item{repos}{character vector of repository urls or links to CITATION.cff
files to be scanned}
}
\value{
A list with CITATION.cff information collected from the
provided repositories
}
\description{
Scan repositories for existing CITATION.cff information
}
\examples{
scanRepos(c("https://github.com/magpiemodel/magpie",
            "https://github.com/remindmodel/remind"))
}
\seealso{
read_cff
}
\author{
Jan Philipp Dietrich
}
