% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff2zenodo.R
\name{cff2zenodo}
\alias{cff2zenodo}
\title{cff2zenodo}
\usage{
cff2zenodo(x)
}
\arguments{
\item{x}{path to citation file or read in citation in citation file format}
}
\description{
Converter citation file format to .zenodo.json metadata file for proper
metadata input to Zenodo.
}
\author{
Jan Philipp Dietrich
}
